Move in-nixpkgs checks to buildenv
